Add files via upload
This is a very basic landing page that shows my implementation of some basic HTML, CSS, and JavaScript. It contains grids, transformations, and a button to toggle the dark theme on and off, and more! I hope to build more complex projects in the future, but this is certainly a great beginning step.